backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
mysql-connector-python==8.2.0
python-dotenv==1.0.0
FastAPI==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
PyJWT==2.8.1
bcrypt==4.1.1
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
TensorFlow==2.15.0
keras==2.15.0
nltk==3.8.1
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
requests==2.31.0
Pillow==10.1.0
opencv-python==4.8.1.78
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
joblib==1.3.2
Werkzeug==3.0.1
pytest==7.4.3
